common/unittest: add some nalreader unit tests
Test the NalReader with some arbitrary data buffers that
contain the start code prefix and some that don't contain
the prefix.

NalReaderTest.ReadAsWhole currently fails since NalReader::read()
always returns true when asWhole==true.  That is, it should
return true only on the first call to read()... assuming data
size > 0.  This needs to be corrected in a future commit.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>